+++ lib86cpu/core/breakpoint.h
@@ -7,8 +7,8 @@
 #pragma once
 
 
-void cpu_check_data_watchpoints(cpu_t *cpu, addr_t addr, size_t size, int type);
-void cpu_check_io_watchpoints(cpu_t *cpu, port_t port, size_t size, int type);
-bool cpu_check_watchpoint_enabled(cpu_t *cpu, int idx);
-int cpu_get_watchpoint_type(cpu_t *cpu, int idx);
-size_t cpu_get_watchpoint_length(cpu_t *cpu, int idx);
+void cpu_check_data_watchpoints(cpu_t *cpu, addr_t addr, uint32_t size, int type);
+void cpu_check_io_watchpoints(cpu_t *cpu, port_t port, uint32_t size, int type);
+bool cpu_check_watchpoint_enabled(cpu_t *cpu, unsigned idx);
+uint32_t cpu_get_watchpoint_type(cpu_t *cpu, unsigned idx);
+uint32_t cpu_get_watchpoint_length(cpu_t *cpu, unsigned idx);
